requirements.txt
ttkbootstrap
mysql-connector-python
pandas